requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# File Processing
python-magic>=0.4.24
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Document Processing
PyPDF2>=2.0.0
python-docx>=0.8.11
openpyxl>=3.0.9
python-pptx>=0.6.21
beautifulsoup4>=4.10.0
lxml>=4.6.3

# Image Processing
Pillow>=9.0.0
torch>=1.9.0
torchvision>=0.10.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.11.0

# User Interface
click>=8.0.0
rich>=10.12.0

# Utilities
tqdm>=4.62.0
python-dateutil>=2.8.2
pytz>=2021.3

# Development
twine>=3.7.1
wheel>=0.37.0
black>=21.12b0
flake8>=4.0.1
pytest>=6.2.5
pytest-cov>=3.0.0
